Rewrite count-rules into ESM format
